Issue 164 benefits news hotfix

+++ packages/benefits/src/controller/DeleteBenefitResolve.js
@@ -5,6 +5,16 @@ module.exports = {
   async DeleteBenefitResolve(request) {
     const { benefit_id } = request.params;
 
+    const benefit = await Benefit.findOne({
+      where: {
+        benefit_id
+      }
+    });
+    
+    if (!benefit){
+      throw { error: "Benefit not found!" };
+    }
+
     await Benefit.destroy({
       where: {
         benefit_id
